refactor: 자판기 클래스 기능별 분리
item manage와 coin recharge로 분리
