+++ src/__test__/coinRecharge.test.ts
@@ -1,7 +1,7 @@
-import VendingMachine from '../ts/VendingMachine';
+import CoinRecharge from '../ts/vendingMachine/CoinRecharge';
 
 describe('잔돈 충전 테스트', () => {
-  const vendingMachine = new VendingMachine();
+  const vendingMachine = new CoinRecharge();
 
   test('자판기 보유 금액을 누적하여 충전할 수 있다.', () => {
     const rechargeCoin = 1000;
